+++ projects/rocprofiler-sdk/source/include/rocprofiler-sdk/hip/api_args.h
@@ -3425,6 +3425,14 @@ typedef union rocprofiler_hip_api_args_t
         const hipLaunchConfig_t* config;
     } hipOccupancyMaxActiveClusters;
 #endif
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 24
+    struct
+    {
+        hipArrayMemoryRequirements* memoryRequirements;
+        hipMipmappedArray_t         mipmap;
+        hipDevice_t                 device;
+    } hipMipmappedArrayGetMemoryRequirements;
+#endif
 } rocprofiler_hip_api_args_t;
-
+ 
 ROCPROFILER_EXTERN_C_FINI