+++ projects/rocprofiler-sdk/source/lib/rocprofiler-sdk/hip/abi.cpp
@@ -628,6 +628,7 @@ ROCP_SDK_ENFORCE_ABI(::HipDispatchTable, hipExtDisableLogging_fn, 508);
 ROCP_SDK_ENFORCE_ABI(::HipDispatchTable, hipExtEnableLogging_fn, 509);
 ROCP_SDK_ENFORCE_ABI(::HipDispatchTable, hipExtSetLoggingParams_fn, 510);
 #endif
+
 #if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 22
 ROCP_SDK_ENFORCE_ABI(::HipDispatchTable, hipMemSetMemPool_fn, 511);
 ROCP_SDK_ENFORCE_ABI(::HipDispatchTable, hipMemGetMemPool_fn, 512);
@@ -636,6 +637,10 @@ ROCP_SDK_ENFORCE_ABI(::HipDispatchTable, hipMemGetMemPool_fn, 512);
 #if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 23
 ROCP_SDK_ENFORCE_ABI(::HipDispatchTable, hipOccupancyMaxPotentialClusterSize_fn, 513);
 ROCP_SDK_ENFORCE_ABI(::HipDispatchTable, hipOccupancyMaxActiveClusters_fn, 514);
+#endif
+ 
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 24
+ROCP_SDK_ENFORCE_ABI(::HipDispatchTable, hipMipmappedArrayGetMemoryRequirements_fn, 515);
 #endif
 
 #if HIP_RUNTIME_API_TABLE_STEP_VERSION == 0
@@ -686,6 +691,8 @@ ROCP_SDK_ENFORCE_ABI_VERSIONING(::HipDispatchTable, 511)
 ROCP_SDK_ENFORCE_ABI_VERSIONING(::HipDispatchTable, 513)
 #elif HIP_RUNTIME_API_TABLE_STEP_VERSION == 23
 ROCP_SDK_ENFORCE_ABI_VERSIONING(::HipDispatchTable, 515)
+#elif HIP_RUNTIME_API_TABLE_STEP_VERSION == 24
+ROCP_SDK_ENFORCE_ABI_VERSIONING(::HipDispatchTable, 516)
 #else
 INTERNAL_CI_ROCP_SDK_ENFORCE_ABI_VERSIONING(::HipDispatchTable, 0)
 #endif